back-end sistema de fila/requirements.txt
Flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.2
python-engineio==4.9.0
mysql-connector-python==8.2.0
python-dotenv==1.0.1